:bug: long to double instead of long to int for getLastSyncDate

+++ android/src/main/java/com/dpppt/DppptModule.java
@@ -58,7 +58,7 @@ public class DppptModule extends ReactContextBaseJavaModule {
         map.putInt("numberOfContacts", tracingStatus.getNumberOfContacts());
         map.putBoolean("isAdvertising", tracingStatus.isAdvertising());
         map.putBoolean("isReceiving", tracingStatus.isReceiving());
-        map.putInt("getLastSyncDate", tracingStatus.getLastSyncDate());
+        map.putDouble("getLastSyncDate", tracingStatus.getLastSyncDate());
         promise.resolve(map);
     }
